Separating out json diff tests from rest
For regular testing we use the json diff utility.
However, this means we cannot test the json diff utility using the json diff utility.
This process is bootstrapped through a custom diffing program and seperate json diff.
It is written in python3 for expediency but should be rewritten in go to reduce the number of languages in this repository.